+++ esp-mdm/src/include/ApplicationBase.h
@@ -20,7 +20,7 @@ class ApplicationBase {
     Callback<void> _process;
     std::string _access_token;
     std::string _authorization;
-    uint32_t _token_expires_at{};
+    int64_t _token_expires_at{};
     std::string _device_name;
     std::string _device_entity_id;
     bool _enable_ota{};
