+++ esp-status-led-support/src/include/WS2812StatusLed.h
@@ -5,8 +5,8 @@
 
 class WS2812StatusLed : public RGBStatusLed, public BrightnessStatusLed {
     led_strip_handle_t _led{};
-    uint32_t _next_update_ms{};  // 0 = disabled, 1 = immediate, else = scheduled time
-    uint32_t _phase_start_ms{};
+    int64_t _next_update_ms{};  // 0 = disabled, 1 = immediate, else = scheduled time
+    int64_t _phase_start_ms{};
     bool _blink_state{};
 
 public:
